Factor task menu out, and make some improvements
