+++ frontend/src/enums/index.js
@@ -1,0 +1,15 @@
+export const TaskRepetitionType = {
+  Single: 'Single',
+  Daily: 'Daily',
+  Weekly: 'Weekly',
+}
+
+export const DaysOfWeek = {
+  Sunday: 'Sun',
+  Monday: 'Mon',
+  Tuesday: 'Tue',
+  Wednesday: 'Wed',
+  Thursday: 'Thu',
+  Friday: 'Fri',
+  Saturday: 'Sat'
+}
