--- frontend/src/features/dashboard/components/DummyComponent.jsx
@@ -1,7 +1,0 @@
-export const DummyComponent = ({propExample}) => {
-  return (
-    <>
-      <p>Hello, I'm a component of {propExample} </p>
-    </>
-  )
-}
